TP2/requirements.txt
pillow==11.0.0
aiohttp==3.10.10